+++ R/conf.mat.R
@@ -13,17 +13,18 @@
 ## - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - |
 
 conf.mat = function( pred, actual, cutoff = NULL, reference = NULL, 
-                     proportion = FALSE, dnn = c( "Prediction", "Actual" ), ... )
+                     proportion = FALSE, dnn = c( "Predict", "Actual" ), ... )
 {
     if( length( pred ) != length( actual ) )
         stop( "prod & actual must have the same length" )
     
-    if( !is.null( cutoff ) ){
-        if( ( cutoff < 0 ) || ( cutoff > 1 ) ) stop( " Value of 'cutoff' must be between 0 and 1." )
+    if( !is.null( cutoff ) )
+    {
+        if( ( cutoff < 0 ) || ( cutoff > 1 ) ) stop( " The value of 'cutoff' must be between 0 and 1." )
         
         levels = base::levels( as.factor( actual ) )
         if( length( levels )  < 2 ) stop( " 'actual' must have more than two levels." )
-        if( length( levels ) != 2 ) stop( " For the case 'cutoff!=NULL', 'actual' must have two levels." )
+        if( length( levels ) != 2 ) stop( " For the case 'cutoff != NULL', 'actual' must have two levels." )
         
         if( levels[ 1 ] == 0 ) levels = c( levels[ 2 ], levels[ 1 ] ) 
         
@@ -40,9 +41,8 @@ conf.mat = function( pred, actual, cutoff = NULL, reference = NULL,
     
     conf_mat = table( pred, actual, dnn = dnn, ... )
     
-    if( proportion == TRUE ){
+    if( proportion == TRUE )
         conf_mat = round( conf_mat / sum( conf_mat ), 3 )
-    }
     
     return( conf_mat )
 }