+++ R/find.na.R
@@ -18,9 +18,10 @@ find.na = function( x )
     
     ind = which( is.na( x ), arr.ind = TRUE )
     
-    if( length( ind ) == 0 ) print( "  There is no missing values (NA)." )
-    
-    return( ind )
+    if( length( ind ) == 0 ) 
+        print( " No missing values (NA) in the dataset." )
+    else
+        return( ind )
 }
    
 ## - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - |